Moved password check to pydash_app.user
